requirements.txt
asgiref==3.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
